requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128

torch==2.7.1+cu128
torchvision==0.22.1+cu128
torchaudio==2.7.1+cu128
ultralytics>=8.0.0
gradio>=6.5.1
opencv-python==4.10.0.84
numpy==1.26.4
loguru>=0.7.0
Pillow>=10.0.0
tqdm>=4.65.0
transformers==4.49.0
click>=8.0.0
einops>=0.8.1
timm>=1.0.24,<1.1.0
accelerate>=1.12.0
supervision>=0.21.0
xformers>=0.0.27.post2
triton-windows<3.7

wheel
huggingface_hub[hf_xet]
setuptools==72.1.0

hydra-core
omegaconf
iopath
ninja
pycocotools
pyyaml
addict
yapf
termcolor
matplotlib

cffi
pycparser
onnx
onnxruntime-gpu

sam2
fairscale